+++ src/App.jsx
@@ -1,18 +1,14 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 import React from "react";
-import ReactDOM from "react-dom";
-import Nav from "./Nav";
-import About from "./About";
-import Intro from "./Intro";
-import Project from "./Project";
-import Blog from "./Blog";
-import Contact from "./Contact";
+import { Navbar } from "./components";
+import { About, Blog, Contact, Intro, Project } from "./pages";
 
 export default function App() {
   return (
     <div>
       <Router>
+      <Navbar />
         <Routes>
           <Route path="/" element={<About />} />
           <Route path="/Intro" element={<Intro />} />
