+++ src/pages/index.js
@@ -1,0 +1,5 @@
+export { default as About } from "./About";
+export { default as Blog } from "./Blog";
+export { default as Intro } from "./Intro";
+export { default as Project } from "./Project";
+export { default as Contact } from "./Contact";
